requirements.txt
revChatGPT
khl.py